+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - [fix:fix reporting bug when port is -1.](https://github.com/Tencent/spring-cloud-tencent/pull/1039)
 - [fix:update guava version.](https://github.com/Tencent/spring-cloud-tencent/pull/1042)
 - [fix:fix circuit breaker bean load order bug when using Nacos discovery.](https://github.com/Tencent/spring-cloud-tencent/pull/1049)
+- [refactor:refactor Polaris registration.](https://github.com/Tencent/spring-cloud-tencent/pull/1056)